Update editing.rst
I tried the column selection procedure as described in the documentation but was unable to see any results. So I experimented with it a lil and found that using my macbook pro trackpad I was able to make the changes I wanted, So I corrected the documentation. Please review and if approved apply it to the original documentation.

I made changes in section 'Column Selection' from line 63 to 65